Complete changes fixed
